refactor for audit crawler WIP
