+++ CxAnalytixDaemon/Daemon.cs
@@ -9,7 +9,7 @@ using System.Reflection;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using CxAnalytix.TransformLogic.Interfaces;
+using CxAnalytix.Interfaces.Outputs;
 
 namespace CxAnalytixDaemon
 {
