+++ CxAnalytixService/ServiceLifecycleControl.cs
@@ -7,7 +7,7 @@ using CxAnalytix.Configuration;
 using System.Reflection;
 using System;
 using CxRestClient;
-using CxAnalytix.TransformLogic.Interfaces;
+using CxAnalytix.Interfaces.Outputs;
 
 [assembly: CxRestClient.IO.NetworkTraceLog()]
 [assembly: log4net.Config.XmlConfigurator(ConfigFile = "CxAnalytixService.log4net", Watch = true)]
