+++ Interfaces/Outputs/IOutput.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace CxAnalytix.TransformLogic.Interfaces
+namespace CxAnalytix.Interfaces.Outputs
 {
     /// <summary>
     /// An interface that is used to output records to some destination
