+++ Log4NetOutput/LoggerOutFactory.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
-using CxAnalytix.TransformLogic.Interfaces;
+using CxAnalytix.Interfaces.Outputs;
 
 namespace CxAnalytix.Out.Log4NetOutput
 {
