+++ TransformLogic/TransformLogic.csproj
@@ -22,6 +22,7 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\CxRestClient\CxRestClient.csproj" />
+		<ProjectReference Include="..\Interfaces\Interfaces.csproj" />
 	</ItemGroup>
 
 </Project>
